feat: split articleList components
