+++ apps/web/components/articleList/index.module.css
@@ -1,0 +1,42 @@
+.list ul {
+  list-style: none;
+  height: 100%;
+}
+.list ul li {
+  border-bottom: 0.02rem solid #272a3a;
+  padding: 0.34rem 0.6rem;
+  height: 2.68rem;
+}
+.list ul li:hover {
+  cursor: pointer;
+}
+.title {
+  color: #ccc8db;
+  font-size: 0.36rem;
+}
+.describe {
+  color: #595c6c;
+  font-size: 0.26rem;
+  margin: 0.06rem 0;
+}
+.detail {
+  color: #ccc8db;
+  font-size: 0.28rem;
+  height: 0.44rem;
+  overflow: hidden;
+}
+.article_info {
+  display: flex;
+  justify-content: flex-start;
+  color: #a2a4b4;
+  font-size: 0.26rem;
+  margin-top: 0.16rem;
+}
+.dianzan {
+  margin-right: 0.4rem;
+  color: #a2a4b4;
+  vertical-align: bottom;
+}
+.dianzan span {
+  vertical-align: bottom;
+}
